jrunak

+++ src/controller/auth/role.controller.js
@@ -22,19 +22,20 @@ const createRole = asyncHandler(async (req, res, next) => {
   }
   const { error, value } = roleValidation.validate(req.body);
   if (error) {
-    const user_id = req.user.id;
     return next(new ErrorResponse(error.details[0].message, 400));
   }
   const test = await getByNameRole(value.name);
   if (test) {
-    const user_id = req.user.id;
     return next(new ErrorResponse("Ushbu ma'lumot avval kiritilgan", 409));
   }
-
+  const name = value.name.trim()
+  if (name === 'region-admin' || name === 'super-admin') {
+    return next(new ErrorResponse("Ushbu rolni kiritish mumkin emas", 400))
+  }
   await create_role(value.name);
   const user_id = req.user.id;
   postLogger.info(`Rol yaratildi: ${value.name}. Foydalanuvchi ID: ${user_id}`);
-  
+
   return res.status(201).json({
     success: true,
     data: "Muvaffaqyatli kiritildi",
@@ -48,10 +49,10 @@ const getAllRole = asyncHandler(async (req, res, next) => {
     return next(new ErrorResponse("Siz uchun ruhsat yo'q", 403));
   }
   const roles = await get_all_role();
-  
+
   const user_id = req.user.id;
   getLogger.info(`Barcha rollar olish muvaffaqiyatli. Foydalanuvchi ID: ${user_id}`);
-  
+
   return res.status(200).json({
     success: true,
     data: roles,
@@ -87,7 +88,7 @@ const updateRole = asyncHandler(async (req, res, next) => {
   await update_role(id, value.name);
   const user_id = req.user.id;
   putLogger.info(`Rol yangilandi: ${value.name}. Foydalanuvchi ID: ${user_id}`);
-  
+
   return res.status(200).json({
     success: true,
     data: "Muvafaqyatli yangilandi",
@@ -105,11 +106,14 @@ const deleteRole = asyncHandler(async (req, res, next) => {
   if (!role) {
     return next(new ErrorResponse("Server xatolik. Malumot topilmadi", 404));
   }
+  if (role.name === 'super-admin' || role.name === 'region-admin') {
+    return next(new ErrorResponse('Ochirib bolmaydi', 400))
+  }
 
   await delete_role(id);
   const user_id = req.user.id;
   deleteLogger.info(`Rol o'chirildi. RoleId: ${id}. Foydalanuvchi ID: ${user_id}`);
-  
+
   return res.status(200).json({
     success: true,
     data: "Muvaffaqiyatli o'chirildi",
